Separated the V3d class into new files More experiments with examples in the test script.
